fix(ipc/window): rename hadlers.ts to handlers.ts and update import
