+++ experimental/beats/src/ipc/window/handlers.ts
